pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["yooshark <yoorudziankou@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.4"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
install = "^1.3.5"
drf-yasg = "^1.21.6"
celery = "^5.3.1"
redis = "^4.6.0"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
mypy = "^1.10.1"
black = "^24.4.2"
pre-commit = "^3.3.3"
pytest-cov = "^4.1.0"
ruff = "^0.0.280"
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
pythonpath = "backend"
testpaths = "tests"
python_files = "tests.py test_*.py *_tests.py"

[tool.black]
line-length = 120
exclude = "(|/.venv|/static/*)/"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_unused_ignores = true
no_warn_no_return = true
show_column_numbers = true
show_error_codes = true
pretty = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 120
exclude = [
    ".git",
    "*/migrations/*",
    "*/static/*",
    ".venv",
    "__init__.py"
]
ignore = ["B904"]
select = ["B","C","E","F","W","T"]

[tool.ruff.mccabe]
max-complexity = 18
